[micro] Further generalize and reorg shrinking code
