+++ dpl_micropolicies/TODO-MP.md
@@ -6,9 +6,11 @@ fold our changes back into master when Nikhil says it's time
 
 * NOW
 
-move Gen into TestHeapSafety
+reorganization
+  the type Color should be defined within the heap safety policy code, not in PIPE.hs
+  see if TestHeapSafety can be removed from imports in Shrinking.hs 
 
-make a second policy
+make a second policy (a trivial one at first)
 
 beautify the code some more
   use the P monad uniformly for printing, I guess
